Implemented Z-buffer
Taking advantage of gamemaker's in built z-buffer through gpu_set_ztestenable, I was able to remove the depth sorting built in to the demo. This optimization increases performance significantly.

In order to make it work I also had to modify the shaders so that empy (alpha == 0) pixels are discarded. If these pixels aren't discarded, they'll be written to the z-buffer, causing parts of the model to dissapear where they are now occluded by the empty pixels.
The z-buffer also fixes a common issue found in sprite stacking games, which is depth sorting with walls, where models frequently end up in front/behind a wall when they shouldn't be. I added a cube object to demonstrate that the system fixes this.